Drop git-build-version build dependency
While it's really nice to have a git-describe derived version reported on
the command line, being forced to use git2 0.3.3 is not so nice. For now,
we're dropping the git-described derived versions, so we can use a newer
git2.

At some point we should really pick up what git-build-version was doing,
but in a way that's compatible with our usage of git2.